+++ approve.php
@@ -315,5 +315,6 @@
 		?>
 	</div>
 </div>
+
 </body>
 </html>