+++ connection.php
@@ -1,6 +1,3 @@
 <?php
-
-	$db=mysqli_connect("localhost","root","","tuitionapp");  
-					// server name, username, password, database name
-
+$db = mysqli_connect("us-cdbr-east-04.cleardb.com", "b13d2a62a9a1fd", "8f70ec7c","heroku_179fff40fae295c");
 ?>