+++ login.php
@@ -81,6 +81,7 @@
     }
 
   ?>
+  
 <?php
 
 include "footer.php";