+++ review.php
@@ -1,4 +1,4 @@
-<?phpp
+<?php
   include "connection.php";
   include "navbar.php";
 ?>
@@ -15,18 +15,18 @@
     <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.3.1/jquery.min.js"></script>
     <script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/js/bootstrap.min.js"></script> 
     <style type="text/css">
-    	/* body
+    	body
     	{
-    		background-image: url("images/66.jpg");
-    		background-repeat: no-repeat;
-    	} */
+    		background-color: black;
+    		
+    	} 
     	.wrapper
     	{
     		padding: 20px;
     		margin: -20px auto;
-    		width:1200px;
-    		height: 900px;
-    		background-color: black;
+    		width:1341px;
+    		height: 350px;
+    		background-color: grey;
     		opacity: .8;
     		color: white;
     	}
@@ -45,19 +45,74 @@
     </style>
 </head>
 <body>
+    
+        <div class="wrapper">
+		<h4>Website Review</h4>
+        <p>If you have any suggesions or questions for the website please comment below.</p>
+        <p> Feel free to leave comments even if you are not a registered user</p>
+
+		<form style="" action="" method="post">
+
+						<div class="form-row">
+							<div class="form-group col-md-4">
+                                <label for="firstname">Name</label>
+                                <input type="text" class="form-control" name="personname" placeholder="Name" required="">
+                            </div>
+                            <div class="form-group col-md-4">
+                                <label for="email">Email</label>
+                                <input type="email" class="form-control" name="email" placeholder="Email" required="">
+                            </div>
+							<div class="form-group col-md-4">
+                                <label for="phone">Contact</label>
+                                <input type="text" 
+                                       onkeypress= "return onlyNumberKey(event)" 
+                                       maxlength="11"class="form-control" name="phone" placeholder="Phone Number" required="">
+                            </div>
+                            
+                        </div>
+
+			<div class="form-row">
+				<input class="form-control" type="text" name="sitecomment" placeholder="Write something..." required=""><br>
+			</div>
+			<input class="btn btn-default" type="submit" name="submit2" value="Comment" style="width: 100px; height: 35px;">		
+		</form>
+		</div>
+	
+	<br>
+<?php
+    if(isset($_POST['submit2'])){
+
+    $reviewername=$_POST['personname'];
+    $email=$_POST['email'];
+    $phone=$_POST['phone'];
+    $sitecomment=$_POST['sitecomment'];
+    
+    //inserts into sitereview in the database with user input
+    $ins="INSERT INTO `sitereview` VALUES('','$reviewername','$email','$phone','$sitecomment');";
+    $result = mysqli_query($db,$ins);
+    ?>
+    <script type="text/javascript">
+           alert("Comment sent successful");
+    </script>
+    <?php
+
+    }
+?>
 
 	<div class="wrapper">
-		<h4>If you have any comments or reviews you want to give to a tutor or tutee, you can do so below.</h4>
-		<br><br><form  style = "color :black" action="" method="post">
+		<h4>Tutor/Student Review.</h4>
+        <p>Only registered users can use this form</p><br>
+        <p>If you have any comments or reviews you want to give to a tutor or tutee, you can do so below.</p>
+		<br><form  style = "color :grey" action="" method="post">
    			<label class="form-label"><b style = "color :white">Username of the tutor/tutee: </b></label>
    			<input class="form-control-lg" type="text" name="targetuser" placeholder="Username" required=""><br>
-   			<br><input class="form-control"  type="text" name="comment" placeholder="Write something..."><br> 
-   			<input class="btn btn-default" type="submit" name="submit" value="submit" style="width: 100px; height: 35px;">  
+   			<br><input class="form-control"  type="text" name="comment" placeholder="Write something..." required=""><br> 
+   			<input class="btn btn-default" type="submit" name="submit1" value="submit" style="width: 100px; height: 35px;">  
   		</form>
 	
 
 		<?php
-			if(isset($_POST['submit']))
+			if(isset($_POST['submit1']))
 			{
 				$targetuser=$_POST['targetuser'];
 				$comment=$_POST['comment'];
@@ -75,6 +130,11 @@
 						}else{
 							$sql="INSERT INTO `review` VALUES ('', '$_SESSION[login_user]','$targetuser','$comment','GOOD') ;";
 							$result = mysqli_query($db,$sql);
+							?>
+							<script type="text/javascript">
+							alert("Your comment has been noted.");
+							</script>
+							<?php
 						}
 				}else{
 					?>
@@ -87,6 +147,21 @@
 		?>
 
 	</div>
+    <br>
+
 	
+	
+
+<?php include "footer.php"; ?>
 </body>
+<script>
+    function onlyNumberKey(evt) {
+
+        // Only ASCII character in that range allowed
+        var ASCIICode = (evt.which) ? evt.which : evt.keyCode
+        if (ASCIICode > 31 && (ASCIICode < 48 || ASCIICode > 57))
+            return false;
+        return true;
+    }
+</script>
 </html>